known_array issue fixed
previously there was an issue where target_array was not equal to the desired output. Target array was longer than the desired output, which means, that known_array had fewer values = 1 than it should have.
The problem was that the way in which I set the pixels of known_array to one was wrong, for further details see the comments in the code.
The unittest now gives 20/20 points. Potential quality improves will take place